pyproject.toml
[project]
name = "atlas-ai-assistant"
version = "0.2.0"
description = "Personal operational AI assistant powered by Claude"
requires-python = ">=3.12,<3.13"

dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.44.0",
    "sqlalchemy>=2.0.49",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.13.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.2.0",
    "google-api-python-client>=2.150.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "anthropic>=0.50.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.0",
    "mypy>=1.20.0",
    "bandit>=1.9.0",
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.0",
    "dotenv-linter>=0.7.0",
    # Windows only: injects system certificates into ssl so requests accepts
    # antivirus/proxy MITM certs. Required locally when SSLError occurs on
    # oauth2.googleapis.com. Has no effect on Linux/macOS.
    "pip-system-certs>=5.0",
]
test = [
    "pytest>=9.0.0",
    "pytest-cov>=7.1.0",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.0",
    "coverage>=7.13.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# ═══════════════════════════════════════════════════════════════════
# RUFF — linter + formatter (replaces flake8 + black + isort)
# ═══════════════════════════════════════════════════════════════════

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["app", "tests"]
exclude = ["scripts/", ".venv/", "htmlcov/"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit (security rules)
    "A",      # flake8-builtins
    "C4",     # flake8-comprehensions
    "T20",    # flake8-print
    "RET",    # flake8-return
    "PTH",    # flake8-use-pathlib
    "RUF",    # ruff-specific rules
]
ignore = [
    "S101",   # assert used (needed in tests)
    "B008",   # function call in default arg (FastAPI Depends pattern)
    "RET504", # unnecessary assignment before return
    "UP007",  # use X | Y for union types (already used)
    "UP017",  # datetime.UTC alias — codebase uses both styles (pre-existing)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # assert is fine in tests
    "S105",   # hardcoded passwords in test fixtures are fine
    "S106",   # same for tokens
]
"app/integrations/**/*.py" = [
    "T201",   # print statements in stubs (if any)
]
"app/core/config.py" = [
    "S105",   # google_token_path is a file path, not a credential
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ═══════════════════════════════════════════════════════════════════
# MYPY — static type checking
# ═══════════════════════════════════════════════════════════════════

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = false
disallow_incomplete_defs = true
no_implicit_optional = true
show_error_codes = true
show_column_numbers = true
pretty = true

plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "app.db.models"
# SQLAlchemy ORM models need relaxed typing
disallow_untyped_defs = false
warn_return_any = false

# ═══════════════════════════════════════════════════════════════════
# PYTEST — test runner + coverage
# ═══════════════════════════════════════════════════════════════════

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-config=pyproject.toml",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = [
    "app/__init__.py",
    "app/integrations/google_mcp.py",
    "app/integrations/rss_reader.py",
]

[tool.coverage.report]
fail_under = 70
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
    "pass",
]
precision = 1

[tool.coverage.html]
directory = "htmlcov"

# ═══════════════════════════════════════════════════════════════════
# BANDIT — security scanner
# ═══════════════════════════════════════════════════════════════════

[tool.bandit]
exclude_dirs = ["tests", "htmlcov", ".venv"]
skips = ["B101"]  # skip assert_used (tests)
